usize->u32

+++ toolkit/paint/Point.zig
@@ -1,19 +1,19 @@
 const Self = @This();
 const std = @import("std");
 
-x: usize,
-y: usize,
+x: u32,
+y: u32,
 
 pub const ZERO = Self{
     .x = 0,
     .y = 0,
 };
 pub const INF = Self{
-    .x = std.math.maxInt(usize),
-    .y = std.math.maxInt(usize),
+    .x = std.math.maxInt(u32),
+    .y = std.math.maxInt(u32),
 };
 
-pub fn init(x: usize, y: usize) Self {
+pub fn init(x: u32, y: u32) Self {
     return Self{ .x = x, .y = y };
 }
